requirements.txt
flask>=3.0
opencv-python-headless>=4.8
inference-sdk>=0.30
python-dotenv>=1.0
numpy>=1.24
scipy>=1.11
